# CORE-207: TSV download optimizations in makeRow()