This PR is an optimization found while researching CORE-207. This should be performance only; it has no functional impact to end users.

The `makeRow()` function is called, as its name implies, for each row of a TSV during download. Thus, it feels worth optimizing.

1. **optimize use of collections within `TSVFormatter.makeRow`**. Previously, we looped over all attributes in the entity, performed a Seq.indexOf() call to determine the column number that attribute should reside in, made a map of the column number->value, and then created a Seq of values based on the map's keys. Now, we loop over the headers in order and extract each column from the entity, still in order. This is less work and makes the `makeRow` function ~1.5x faster.
2. **use List instead of IndexedSeq when iterating**. We were using `IndexedSeq` all over the place. `IndexedSeq` is optimized more for random access, while `List` (which extends `LinearSeq`) is optimized more for sequential access. Our access was sequential. Changing this added another ~1.4x speedup, for a total of ~2.2x.

Raw numbers from the JMH benchmark:
  | ops/sec | error +/- | vs.baseline | avg ops/sec | avg vs. baseline
-- | -- | -- | -- | -- | --
baseline | 244,195.39 | 19,087.54 |   | 247,734.34 |  
  | 240,787.36 | 21,066.81 | 98.60% |   |  
  | 258,220.27 | 14,516.10 | 105.74% |   |  
Map->Seq for makerow | 388,638.98 | 48,430.52 | 159.15% | 398,863.48 | 161.00%
  | 375,149.40 | 43,361.68 | 153.63% |   |  
  | 325,224.09 | 45,584.79 | 133.18% |   |  
  | 315,577.37 | 51,806.07 | 129.23% |   |  
  | 410,244.36 | 38,258.30 | 168.00% |   |  
IndexedSeq -> List | 578,346.66 | 8,294.20 | 236.84% | 555,535.68 | 224.25%
  | 559,476.43 | 72,216.37 | 229.11% |   |  
  | 546,860.46 | 24,416.72 | 223.94% |   |  
  | 537,459.14 | 25,614.79 | 220.09% |   |  


---

Have you read [CONTRIBUTING.md](../CONTRIBUTING.md) lately? If not, do that first.

I, the developer opening this PR, do solemnly pinky swear that:

- [x] I've followed [the instructions](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#api-changes) if I've made any changes to the API, _especially_ if they're breaking changes
- [x] I've updated the [FISMA documentation](https://github.com/broadinstitute/firecloud-orchestration/blob/develop/CONTRIBUTING.md#fisma-documentation-changes) if I've made any security-related changes, including auth, encryption, or auditing

In all cases:

- [ ] Get two thumbsworth of review and PO signoff if necessary
- [x] Verify all tests go green
- [ ] Squash and merge. Make sure your branch deletes; GitHub should do this for you.
- [ ] Test this change deployed correctly and works on dev environment after deployment